src/requirements.txt
git+https://github.com/Rapptz/discord.py
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==3.0.1
requests==2.22.0
beautifulsoup4==4.9.0
python-dateutil==2.8.0
Pillow==6.1.0
redis==3.5.2